Initial commit of the coding bee web app!
Adds a Node.js server which opens a web socket. Clients that connect send 'keystroke' events containing a keyCode and the server appends it to a list.

Adds some TypeScript client code which listens on window keydown events and sends a 'keystroke' event to the server via the web socket.